DEV-1 improve file structure
