+++ src/Shared/PacketAdapter/infra/tests/resources/empty.pcapng
