+++ src/Shared/PacketAdapter/infra/tests/resources/mixed.pcapng
