+++ src/Shared/PacketAdapter/infra/tests/resources/noise.txt
