+++ src/Shared/PacketAdapter/infra/transportController.go
@@ -1,25 +1,31 @@
 package PacketAdapter
 
-import "net"
+import (
+	"net"
+
+	"github.com/HyperloopUPV-H8/Backend-H8/Shared/PacketAdapter/infra/aliases"
+	"github.com/HyperloopUPV-H8/Backend-H8/Shared/PacketAdapter/infra/sniffer"
+	"github.com/HyperloopUPV-H8/Backend-H8/Shared/PacketAdapter/infra/tcp"
+)
 
 type TransportController struct {
-	sniffer Sniffer
-	server  Server
+	sniffer sniffer.Sniffer
+	server  tcp.Server
 }
 
 var (
 	snifferTarget string = "\\Device\\NPF_Loopback"
 	snifferLive   bool   = true
 
-	serverPort Port = 6100
+	serverPort aliases.Port = 6000
 )
 
 func NewTransportController(validAddrs []string) TransportController {
-	validAddrIPs := stringsToIPs(validAddrs)
+	validAddrIPs := aliases.StringsToIPs(validAddrs)
 
 	return TransportController{
-		sniffer: NewSniffer(snifferTarget, snifferLive, createFilters(validAddrIPs)),
-		server:  OpenServer(serverPort, validAddrIPs),
+		sniffer: sniffer.New(snifferTarget, snifferLive, createFilters(validAddrIPs)),
+		server:  tcp.Open(serverPort, validAddrIPs),
 	}
 }
 
@@ -28,37 +34,37 @@ func (controller TransportController) ReceiveData() []byte {
 }
 
 func (controller TransportController) ReceiveMessage() [][]byte {
-	return payloadsToBytes(controller.server.ReceiveNext())
+	return aliases.PayloadsToBytes(controller.server.ReceiveNext())
 }
 
 func (controller TransportController) Send(addr string, payload []byte) {
-	controller.server.Send(IP(addr), payload)
+	controller.server.Send(aliases.IP(addr), payload)
 }
 
 func (controller TransportController) AliveConnections() []string {
-	return ipsToStrings(controller.server.ConnectedAddresses())
+	return aliases.IPsToStrings(controller.server.ConnectedAddresses())
 }
 
 func (controller TransportController) Close() {
 	controller.server.Close()
 }
 
-func createFilters(validAddrIPs []IP) []Filterer {
+func createFilters(validAddrIPs []aliases.IP) []sniffer.Filterer {
 	ipRange := append(validAddrIPs, getLocalIPs()...)
-	return []Filterer{UDPFilter{}, SourceIPFilter{validAddrIPs}, DestinationIPFilter{ipRange}}
+	return []sniffer.Filterer{sniffer.UDPFilter{}, sniffer.SourceIPFilter{SrcIPs: validAddrIPs}, sniffer.DestinationIPFilter{DstIPs: ipRange}}
 }
 
-func getLocalIPs() []IP {
+func getLocalIPs() []aliases.IP {
 	ifaces, err := net.Interfaces()
 	if err != nil {
 		return nil
 	}
 
-	ips := make([]IP, 0, len(ifaces))
+	ips := make([]aliases.IP, 0, len(ifaces))
 	for _, iface := range ifaces {
 		addrs, _ := iface.Addrs()
 		for _, addr := range addrs {
-			ips = append(ips, IP(addr.String()))
+			ips = append(ips, aliases.IP(addr.String()))
 		}
 	}
